backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
opencv-python
pillow
python-dotenv